airflow/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-google
pyarrow